Did some refactorng and physics finally work

+++ src/main/java/components/GridLines.java
@@ -14,21 +14,21 @@ public class GridLines extends Component {
         Vector2f caperaPosition = camera.position;
         Vector2f projectionSize = camera.getProjectionSize();
 
-        int firstX = (int) ((caperaPosition.x / Settings.GRID_WIDTH - 1)) * Settings.GRID_WIDTH;
-        int firstY = (int) ((caperaPosition.y / Settings.GRID_HEIGHT - 1)) * Settings.GRID_HEIGHT;
+        float firstX = (caperaPosition.x / Settings.GRID_WIDTH - 1) * Settings.GRID_WIDTH;
+        float firstY = (caperaPosition.y / Settings.GRID_HEIGHT - 1) * Settings.GRID_HEIGHT;
 
         int numVerLines = (int) (projectionSize.x * camera.getZoom() / Settings.GRID_WIDTH) + 2;
         int numHorLines = (int) (projectionSize.y * camera.getZoom() / Settings.GRID_HEIGHT) + 2;
 
-        int width = (int) (projectionSize.x * camera.getZoom()) + Settings.GRID_WIDTH * 2;
-        int height = (int) (projectionSize.y * camera.getZoom()) + Settings.GRID_HEIGHT * 2;
+        float width = projectionSize.x * camera.getZoom() + Settings.GRID_WIDTH * 2;
+        float height = projectionSize.y * camera.getZoom() + Settings.GRID_HEIGHT * 2;
 
         int maxLines = Math.max(numVerLines, numHorLines);
         Vector3f color = new Vector3f(0.2f, 0.2f, 0.2f);
 
         for (int i = 0; i < maxLines; i++) {
-            int x = firstX + (Settings.GRID_WIDTH * i);
-            int y = firstY + (Settings.GRID_HEIGHT * i);
+            float x = firstX + (Settings.GRID_WIDTH * i);
+            float y = firstY + (Settings.GRID_HEIGHT * i);
 
             if (i < numVerLines) {
                 DebugDraw.addLine2D(new Vector2f(x, firstY), new Vector2f(x, firstY + height), color, 2);
